backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
PyJWT==2.8.0
Pillow==10.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
